Fix compile issue

+++ src/easylogging++.cc
@@ -1116,7 +1116,7 @@ base::type::string_t DateTime::formatTime(unsigned long long time, base::Timesta
     if (base::consts::kTimeFormats[i].value == 1000.0f && time / 1000.0f < 1.9f) {
       break;
     }
-    time /= base::consts::kTimeFormats[i].value;
+    time /= (uint64_t)base::consts::kTimeFormats[i].value;
     unit = base::consts::kTimeFormats[i + 1].unit;
   }
   base::type::stringstream_t ss;
